requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
discord-components==2.1.2
discord-ui==5.1.6
discord.py==1.7.3
emoji-country-flag==1.3.0
ffmpeg-python==0.2.0
future==0.18.2
fuzzywuzzy==0.18.0
idna==3.3
imageio-ffmpeg==0.4.6
multidict==6.0.2
numpy==1.22.3
py-cord==2.0.0b7
pycountry==22.3.5
python-dotenv==0.20.0
python-Levenshtein==0.12.2
requests==2.27.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
